Updated Project Structure
Each model holds all code related to the model.
