+++ .gitignore
@@ -10,3 +10,6 @@ target/
 /.idea/
 CMakeLists.txt
 cmake-build-debug/
+
+*.env
+*.sqlite
