+++ Cargo.toml
@@ -7,7 +7,7 @@ authors = ["Mcat12 <newtoncat12@yahoo.com>", "0xcaff <martincharles07@gmail.com>
 lto = true
 
 [dependencies]
-diesel = { version = "1.0.0", features = ["postgres"] }
+diesel = { version = "1.0.0", features = ["postgres", "sqlite"] }
 dotenv = "0.9.0"
 iron = "0.6.0"
 mount = "0.4.0"
@@ -15,11 +15,7 @@ logger = "0.4.0"
 persistent = "0.4.0"
 juniper = "0.9.2"
 juniper_iron = "0.1.2"
-redis = "0.8.0"
 r2d2 = "0.8.0"
 r2d2-diesel = "1.0.0"
-serde = "1.0"
-serde_derive = "1.0"
-serde-redis = "0.7.1"
 chrono = "0.4"
 taglib2-sys = { path = "taglib2-sys" }
