+++ src/context.rs
@@ -1,24 +1,21 @@
-extern crate diesel;
+extern crate dotenv;
 extern crate r2d2;
 extern crate r2d2_diesel;
-extern crate dotenv;
 
-use self::diesel::pg::PgConnection;
-
-use std::ops::Deref;
-use std::error::Error;
+use diesel::sqlite::SqliteConnection;
 use iron::prelude::*;
 use iron::typemap::Key;
 use juniper;
 use persistent::Read;
-use iron::BeforeMiddleware;
-use iron::AfterMiddleware;
+use std;
+use std::error::Error;
 
-pub type ConnectionManager = r2d2_diesel::ConnectionManager<PgConnection>;
-pub type Pool = r2d2::Pool<ConnectionManger>;
+pub type ConnectionManager = r2d2_diesel::ConnectionManager<SqliteConnection>;
+pub type Pool = r2d2::Pool<ConnectionManager>;
+pub type PooledConnection = r2d2::PooledConnection<ConnectionManager>;
 
 pub struct IronContext {
-    pub connection_manager: r2d2::PolledConnectionManager<PgConnection>,
+    pub connection_manager: Pool,
 }
 
 #[derive(Copy, Clone)]
@@ -28,28 +25,30 @@ impl Key for ContextKey {
 }
 
 impl IronContext {
-    pub fn init_middleware() -> Result<(BeforeMiddleware, AfterMiddleware), Box<Error>> {
+    pub fn init_middleware() -> Result<(Read<ContextKey>, Read<ContextKey>), Box<Error>> {
         let database_url = dotenv::var("DATABASE_URL")?;
-        let manager = ConnectionManager::new(database_url)?;
-        let pool: Pool = r2d2::Pool(manager)?;
+        let manager = ConnectionManager::new(database_url);
+        let pool: Pool = r2d2::Pool::new(manager)?;
+        let context = IronContext {
+            connection_manager: pool,
+        };
 
-        Read::<database::ConnectionKey>::both(pool)
+        Ok(Read::<ContextKey>::both(context))
     }
 }
 
-pub struct GraphQLContext {
-    pub connection: r2d2::PooledConnection<r2d2_diesel::ConnectionManager<PgConnection>>,
+unsafe impl std::marker::Sync for IronContext {}
 
+pub struct GraphQLContext {
+    pub connection: PooledConnection,
 }
 
 impl GraphQLContext {
     pub fn from_request(request: &mut Request) -> Self {
         let iron_context = request.get::<Read<ContextKey>>().unwrap();
-        let connection = iron_context.connection_manager.get();
+        let connection: PooledConnection = iron_context.connection_manager.get().unwrap();
 
-        GraphQLContext {
-            connection,
-        }
+        GraphQLContext { connection }
     }
 }
