+++ src/lib.rs
@@ -1,17 +1,17 @@
 extern crate chrono;
 extern crate iron;
+
 #[macro_use]
 extern crate juniper;
 extern crate juniper_iron;
 extern crate logger;
 extern crate mount;
 extern crate persistent;
-extern crate redis;
-extern crate serde;
-#[macro_use]
-extern crate serde_derive;
-extern crate serde_redis;
 
-pub mod schema;
+#[macro_use]
+extern crate diesel;
+
 pub mod context;
+pub mod database;
+pub mod models;
 pub mod server;
