--- src/schema/mod.rs
@@ -1,3 +1,0 @@
-pub mod model;
-mod binding;
-mod resolvers;
