+++ src/server.rs
@@ -1,26 +1,25 @@
 extern crate dotenv;
 
-use std::env;
-use juniper_iron::{GraphQLHandler, GraphiQLHandler};
+use context::{GraphQLContext, IronContext};
 use iron::{Chain, Iron};
-use mount::Mount;
+use juniper_iron::{GraphQLHandler, GraphiQLHandler};
 use logger::Logger;
-use schema::model::{Mutation, Query};
-use context;
+use models::{Mutation, Query};
+use mount::Mount;
 
 pub fn start() {
     let mut mount = Mount::new();
-    let mut chain = Chain::new(mount);
 
     // Register Routes
-    let graphql_handler = GraphQLHandler::new(context::IronContext::from_request, Query, Mutation);
+    let graphql_handler = GraphQLHandler::new(GraphQLContext::from_request, Query, Mutation);
     mount.mount("/graphql", graphql_handler);
 
     let graphiql_handler = GraphiQLHandler::new("/graphql");
     mount.mount("/", graphiql_handler);
 
     // Setup Context Middleware
-    chain.link(context::init_context_middleware().expect("failed to initialize context"));
+    let mut chain = Chain::new(mount);
+    chain.link(IronContext::init_middleware().expect("failed to initialize context"));
 
     // Setup Logging Middleware
     let (logger_before, logger_after) = Logger::new(None);
@@ -28,8 +27,8 @@ pub fn start() {
     chain.link_after(logger_after);
 
     // Start Server
-    let host = dotenv::var("DATABASE_URL").map(|s| { s.as_str() }).unwrap_or("0.0.0.0:8080");
+    let host = dotenv::var("HOST").unwrap_or_else(|_| "0.0.0.0:8080".to_owned());
     println!("Starting Server on {}", host);
     let iron = Iron::new(chain);
-    iron.http(host).unwrap();
+    iron.http(host.as_str()).unwrap();
 }
